requirements.txt
requests~=2.31.0
telebot~=0.0.5
peewee~=3.17.1
python-dotenv~=1.0.1
pydantic~=2.6.4